Add timer length configuration for shop

+++ Assets/Scripts/Game/CountdownTimer.cs
@@ -3,15 +3,15 @@ using System.Collections;
 using UnityEngine;
 
 public class CountdownTimer : MonoBehaviour {
-    private int countdownDurationInSeconds;
-    private int timeRemaining;
+    private float countdownDurationInSeconds;
+    private float timeRemaining;
     private float speedMultiplier = 1f;
     private const float MinimumSpeedMultiplier = 0.1f;
     public Action OnTimerEnd;
     public Action<int> OnTick;
     public Action OnReset;
 
-    public void StartTimer(int ShopDurationInSeconds) {
+    public void StartTimer(float ShopDurationInSeconds) {
         this.countdownDurationInSeconds = ShopDurationInSeconds;
         StartCoroutine(Countdown());
     }
@@ -19,7 +19,7 @@ public class CountdownTimer : MonoBehaviour {
     private IEnumerator Countdown() {
         timeRemaining = countdownDurationInSeconds;
         while (timeRemaining > 0) {
-            OnTick?.Invoke(timeRemaining);
+            OnTick?.Invoke((int)timeRemaining);
             yield return new WaitForSeconds(1f/speedMultiplier);
             timeRemaining--;
         }
